Store RSS benchmark with baseline subtracted
The baseline RSS memory changes depending on the environment (the Julia
sysimage, but apparently also the Python version). So if we run
consecutive benchmarks, the baseline RSS might not be the same for all
of the resulting data. Thus, we now directly subtract the baseline RSS
from the data before storing it.

The baseline is stored in the summary data file, so that the original
RSS value can be recovered if necessary.