Use grad-via-U for Semi-AD PE/C benchmarks
Before, the Semi-AD benchmarks for PE and C were using the the gradient
via χ. Now they use the gradient via U, in line with the text in the
paper. This is because if we use τ for the SM-optimization, it makes
more sense to use U for PE/C: in both cases, we're using the Semi-AD
with the lowest complexity. It doesn't really make a noticeable
difference in the resulting plot, but it's conceptually cleaner.